refactor: DataProvider protocol + ProxmoxProvider facade (v3.5 groundwork)
- new provider/_provider.py: runtime_checkable DataProvider protocol
  (nodes/vms/cluster/storage/tasks/pools/access/rrd + close, context
  manager) and ProxmoxProvider facade lazily wrapping ProxmoxSession
  and the typed XxxAPI classes
- backend.py depends on the facade instead of constructing
  ProxmoxSession + 8 API classes per worker (59 session sites,
  64 ctor sites); _await_task now takes the provider and uses
  provider.tasks.poll
- ui/api/metrics.py uses ProxmoxProvider.rrd; raw requests.Session
  download paths untouched
- tests: shared FakeProvider stub, patch backend.ProxmoxProvider
- docs: ARCHITECTURE.md 'Data Provider (realization)' section,
  ROADMAP v3.5 Data Provider API marked done
- ProxmoxSession timeout typed float (facade passes float through)

+++ pve_center/provider/_session.py
@@ -49,7 +49,7 @@ def _q(value) -> str:
 class ProxmoxSession:
     """Wraps ProxmoxAPI with unified error handling and connection lifecycle."""
 
-    def __init__(self, cfg: dict, timeout: int = 15) -> None:
+    def __init__(self, cfg: dict, timeout: float = 15) -> None:
         self.cfg = cfg
         self.timeout = timeout
         self._proxmox: ProxmoxAPI | None = None
